Refactor: application dto로 통합한다
