+++ src/types/dto/index.ts
@@ -2,5 +2,4 @@ export * from './adminMember';
 export * from './base';
 export * from './application';
 export * from './applicationForm';
-export * from './applicationById';
 export * from './team';
